refactor: split orchestrator module
- extract config, runtime, engine, traits, and types into dedicated files

- update whisper setup, runtime handle helpers, and tests to avoid dead code warnings

- refresh core orchestrator refactor design notes and mark task 1.0 complete
