+++ core/src/orchestrator/runtime/util.rs
@@ -1,0 +1,10 @@
+use std::time::Duration;
+
+pub(crate) fn duration_to_ms(duration: Duration) -> u64 {
+    let millis = duration.as_millis();
+    if millis > u128::from(u64::MAX) {
+        u64::MAX
+    } else {
+        millis as u64
+    }
+}
