+++ core/src/orchestrator/tests/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+mod runtime;
