+++ tasks/tasks-0007-large-files-refactor.md
@@ -1,0 +1,18 @@
+# 大文件拆分实施任务
+
+## 任务分解
+- [x] 1.0 `core/src/orchestrator/mod.rs` 拆分
+- [ ] 2.0 `apps/desktop/src/App.tsx` 拆分
+- [ ] 3.0 `core/src/session/mod.rs` 拆分
+- [ ] 4.0 `apps/desktop/src-tauri/src/audio.rs` 拆分
+- [ ] 5.0 `apps/desktop/src-tauri/src/hotkey.rs` 拆分
+- [ ] 6.0 `apps/desktop/src/features/transcription/DualViewPanel.tsx` 拆分
+- [ ] 7.0 `apps/desktop/src-tauri/src/native_probe.rs` 拆分
+- [ ] 8.0 `apps/desktop/src-tauri/src/main.rs` 拆分
+- [ ] 9.0 `apps/desktop/src/features/transcription/hooks/useDualViewTranscript.ts` 拆分
+- [x] 10.0 `core/src/session/publisher.rs` 拆分
+- [ ] 11.0 `apps/desktop/src-tauri/src/session.rs` 拆分
+
+## 相关文件
+- docs/large-file-refactor-plan.md — 原始拆分方案
+- tasks/tasks-0007-large-files-refactor.md — 当前任务追踪
